third_party/blink/web_tests/platform/generic/fast/block/margin-left-margin-right-auto-expected.txt

PASS rtl_ml_in_ltr.getBoundingClientRect().right is 700
PASS ltr_ml_in_ltr.getBoundingClientRect().right is 700
PASS rtl_mr_in_ltr.getBoundingClientRect().left is 0
PASS ltr_mr_in_ltr.getBoundingClientRect().left is 0
PASS rtl_ml_in_rtl.getBoundingClientRect().right is 700
PASS ltr_ml_in_rtl.getBoundingClientRect().right is 700
PASS rtl_mr_in_rtl.getBoundingClientRect().left is 0
PASS ltr_mr_in_rtl.getBoundingClientRect().left is 0
PASS rtl_ml_in_ltr_r.getBoundingClientRect().right is 700
PASS ltr_ml_in_ltr_r.getBoundingClientRect().right is 700
PASS rtl_mr_in_ltr_r.getBoundingClientRect().left is 0
PASS ltr_mr_in_ltr_r.getBoundingClientRect().left is 0
PASS rtl_ml_in_rtl_l.getBoundingClientRect().right is 700
PASS ltr_ml_in_rtl_l.getBoundingClientRect().right is 700
PASS rtl_mr_in_rtl_l.getBoundingClientRect().left is 0
PASS ltr_mr_in_rtl_l.getBoundingClientRect().left is 0
PASS successfullyParsed is true

TEST COMPLETE
